Add recipes refactor into small components
