+++ src/components/recipes/addrecipe/AddRecipe.css
@@ -1,49 +1,3 @@
-.recipe_ingredients {
-  max-height: 450px;
-  overflow: hidden;
-  min-width: 200px;
-  flex-grow: 2;
-  overflow-y: auto;
-  min-height: 300px;
-  max-width: 500px;
-}
-
-.recipe_ingredients_table {
-  max-height: 450px;
-  min-height: 300px;
-}
-
-.ingredients_options {
-  max-height: 300px;
-  overflow: hidden;
-  min-width: 50px;
-  margin-bottom: 30px;
-  min-height: 200px;
-  max-width: 300px;
-}
-
-.ingredient_options_container {
-  display: flex;
-  flex-direction: column;
-  flex-grow: 1;
-  min-height: 200px;
-}
-
-.ingredients_options_table {
-  max-height: 300px;
-}
-
-.new_ingredient_textfield {
-  flex-grow: 1;
-  flex-shrink: 0;
-}
-
-.new_ingredient_span {
-  flex-grow: 2;
-  flex-shrink: 0;
-  min-width: 5px;
-}
-
 .recipe_name {
   width: 30%;
   margin: auto;
